fw_iso_resource: change signal prototype to receive instance of GError
Current implementation returns raw code of errno to report error for
allocation/deallocation requests. This is not necessarily convenient
to developers and users for fixing bug and recoverying error.

This commit changes current prototype of allocated/deallocated GObject
signals to receive an instance of GError for error reporting. The
change breaks backward compatibility.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>