fw_iso_ctx: allow function callers to schedule hardware interrupt
In current design of libhinoko, schedule of hardware interrupt is
transparent from application. However, it's convenient for some
applications to schedule hardware interrupt. For example, audio
application typically process the same amount of audio data frame
at once.

As a first step, this commit adds local function for the purpose. In
this time, schedule of hardware interrupt is still transparent.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>