fw_iso_rx_multiple: allow application to configure 0 for interrupt interval
Unlike Hinoko.FwIsoTx and Hinoko.FwIsoRxSingle, the context used by
Hinoko.FwIsoRxMultiple (FW_CDEV_ISO_CONTEXT_RECEIVE_MULTICHANNEL)
doesn't queue events per packet. It splits buffer shared with
application for several parts, then fill the part with the content of
packet. When one of the parts is filled, then queue event when scheduled
for hardware interrupt. The context is not suitable to queue the event
per packet.

Due to the above design, it's not so convenient for application to
schedule hardware interrupt voluntarily. On the other hand, the control
graph with flushing completed packets is still convenient in the case.

This commit allows application to suppress schedule of hardware
interrupt. The application should call
Hinoko.FwIsoCtx.flush_completions() when suppress it.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>