fw_iso_resource: rewrite public API according to GNOME convention
In GNOME convention for throw function which has an argument of GError to
report failure, the function is programmed to return value of gboolean
type to report whether the argument is filled with instance of GError.

This commit rewrites public APIs according to it. The signature of these
APIs are changed without backward compatibility.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>