enums: rename mode of isochronous contexts
In teminology of device driver development, the choice of word is
device-oriented, thus 'rx' is for incoming data path to target device,
and 'tx' is for outgoing data path from target device.

Current implementation uses the words for different ways. This commit
corrects the usage of words with loss of backward compatibility.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>