fw_iso_resource_once: remove async/sync suffixes from deallocation methods
The request of deallocation generates completion event. In current
implementation, the method for request has async suffix, and the method to
wait for the event has sync suffix. However, these two keywords are often
heavy use in different contexts and sometimes unfriendly to developers and
users.

This commit renames these methods. The name of async method becomes
simplier, and the sync method is suffixed with wait.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>