fw_iso_resource: remove sync suffix from allocation methods
The request of allocation generates completion event. In current
implementation, the method for request has async suffix, and the method to
wait for the event has sync suffix. However, these two keywords are often
heavy use in different contexts and sometimes unfriendly to developers and
users.

This commit renames sync method with wait suffix.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>